cephalopy_package/pyproject.toml
[tool.poetry]
name = "cephalopy"
version = "0.1.0"
description = "A python package for automatic landmark detection and cephalometric analysis"
authors = ["VictorOorthuis <233189@buas.nl>"]
readme = "README.md"
packages = [{include = "cephalopy", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10.11,<3.15"
numpy = ">=1.26.0"
scikit-learn = ">=1.5.0"
tqdm = ">=4.66.0"
matplotlib = ">=3.9.0"
pillow = ">=10.4.0"
opencv-python = "^4.10.0.88"
scikit-image = "^0.24.0"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
python-multipart = "^0.0.20"
reportlab = "^4.2.4"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0"
pytest-cov = ">=5.0.0"
pytest-asyncio = ">=0.24.0"
pytest-xdist = ">=3.6.0"  # For parallel test execution
pre-commit = ">=3.8.0"
ruff = ">=0.5.7"
ghp-import = ">=2.1.0"
poetry2conda = ">=0.3.0"
codecov = "^2.1.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
preview = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-vv",
    "--cov=cephalopy",
    "--cov=api",
    "--cov-report=term-missing",
    "--tb=short",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
pythonpath = ["src"]
